Fix: spot register 페이지 태그 중복 코드 제거
