+++ src/components/commons/BadgeListsController/index.tsx
@@ -1,5 +1,5 @@
+import BadgeLists from "@/components/commons/BadgeList/BadgeLists";
 import { Control, Controller, FieldValues, Path } from "react-hook-form";
-import BadgeLists from "../spotRegister/BadgeLists";
 
 interface BadgeListsControllerProps<ControlType extends FieldValues> {
   formFieldName: Path<ControlType>;
@@ -15,7 +15,7 @@ const BadgeListsController = <ControlType extends FieldValues>({
       control={control}
       name={formFieldName}
       render={({ field }) => {
-        return <BadgeLists selectedDestinationBadges={field.value} onChange={field.onChange} />;
+        return <BadgeLists selectedBadges={field.value} onChange={field.onChange} />;
       }}
     />
   );
